requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.5.0
# Vision Framework для macOS (заменяет Tesseract)
pyobjc-framework-Vision>=12.0
Pillow>=10.2.0
opencv-python>=4.8.0
pdf2image==1.16.3
python-docx==1.1.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
pymongo==4.6.0

# ML зависимости для Фазы 2
torch>=2.0.0
transformers>=4.35.0
torchvision>=0.15.0
sentencepiece>=0.1.99  # Для T5 токенизаторов
# tensorflow>=2.13.0  # Временно отключен из-за проблем с Python 3.13
# keras>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
